Move PageXray to own plugin (#1877)
* Move PageXray to own plugin.

This makes it easier for WebPageTest (and other tools in the future)
to xray the HAR file.
