+++ lib/core/pluginLoader.js
@@ -9,6 +9,7 @@ Promise.promisifyAll(fs);
 const defaultPlugins = new Set([
   'browsertime',
   'coach',
+  'pagexray',
   'domains',
   'assets',
   'html',
