+++ lib/plugins/coach/index.js
@@ -1,7 +1,5 @@
 'use strict';
 const aggregator = require('./aggregator');
-const pagexrayAggregator = require('./pagexrayAggregator');
-const pagexray = require('pagexray');
 
 const DEFAULT_METRICS_SUMMARY = [
   'score.*',
@@ -21,32 +19,6 @@ const DEFAULT_METRICS_PAGESUMMARY = [
   'advice.info.scripts',
   'advice.info.localStorageSize'
 ];
-const DEFAULT_PAGEXRAY_PAGESUMMARY_METRICS = [
-  'contentTypes',
-  'transferSize',
-  'contentSize',
-  'requests',
-  'firstParty',
-  'thirdParty',
-  'responseCodes',
-  'expireStats',
-  'totalDomains',
-  'lastModifiedStats',
-  'cookieStats'
-];
-const DEFAULT_PAGEXRAY_SUMMARY_METRICS = [
-  'contentTypes',
-  'transferSize',
-  'contentSize',
-  'requests',
-  'firstParty',
-  'thirdParty',
-  'responseCodes',
-  'expireStats',
-  'domains',
-  'lastModifiedStats',
-  'cookieStats'
-];
 
 module.exports = {
   open(context, options) {
@@ -60,14 +32,6 @@ module.exports = {
     context.filterRegistry.registerFilterForType(
       DEFAULT_METRICS_PAGESUMMARY,
       'coach.pageSummary'
-    );
-    context.filterRegistry.registerFilterForType(
-      DEFAULT_PAGEXRAY_PAGESUMMARY_METRICS,
-      'pagexray.pageSummary'
-    );
-    context.filterRegistry.registerFilterForType(
-      DEFAULT_PAGEXRAY_SUMMARY_METRICS,
-      'pagexray.summary'
     );
   },
   processMessage(message, queue) {
@@ -92,46 +56,12 @@ module.exports = {
         break;
       }
 
-      case 'browsertime.har': {
-        const url = message.url;
-        const group = message.group;
-        let config = {
-          includeAssets: true,
-          firstParty: this.options.firstParty
-            ? this.options.firstParty
-            : undefined
-        };
-        const pageSummary = pagexray.convert(message.data, config);
-
-        pagexrayAggregator.addToAggregate(pageSummary, group);
-
-        queue.postMessage(
-          make('pagexray.pageSummary', pageSummary[0], { url, group })
-        );
-
-        pageSummary.forEach((run, runIndex) => {
-          queue.postMessage(
-            make('pagexray.run', run, { url, group, runIndex })
-          );
-        });
-        break;
-      }
-
       case 'sitespeedio.summarize': {
         let summary = aggregator.summarize();
         if (summary) {
           for (let group of Object.keys(summary.groups)) {
             queue.postMessage(
               make('coach.summary', summary.groups[group], { group })
-            );
-          }
-        }
-
-        let pagexraySummary = pagexrayAggregator.summarize();
-        if (pagexraySummary) {
-          for (let group of Object.keys(pagexraySummary.groups)) {
-            queue.postMessage(
-              make('pagexray.summary', pagexraySummary.groups[group], { group })
             );
           }
         }
