+++ lib/plugins/pagexray/pagexrayAggregator.js
